Modified the train script such that reporting works correctly with tensorboard changed the number of output nodes to reflect the labels (there are 78 lables, but the highest label is 91, so we need 91 nodes)
using the example code from tensorflow to get the reporting, and got the layer class etc for free. had to get rid of the model function as it was instantiating multiple calls for summary variables (once for test and once for train).

using the feed dict funciton too which is pretty cool. you can use it to either feed the network training data, or test data.